refactor: move notifier to service, split player.rs by topic
Notifier doesn't play tracks — it manages timed reminders — so it
belongs alongside break_service and gather_service in service/, not in
player/. Renamed to notifier_service.rs to match the suffix used by the
rest of service/.

Split player.rs (656 lines) into player.rs (the Player state machine
plus activity/normalization helpers) and track.rs (the data types:
Track, Playlist, TrackSource, TrackMetadata, PlaybackError) so each
file has one concern. CLAUDE.md updated to spell out that player/ is
only for the playback engine.
