+++ src/bot.rs
@@ -3,8 +3,9 @@ use crate::commands::{activity, music, reputation, utility};
 use crate::service::break_service::BreakState;
 use crate::service::gather_service::GatherState;
 use crate::handlers::{error_handler, voice_handler};
-use crate::player::notifier::{Notifier, NotifierError};
-use crate::player::player::{PlaybackError, Player};
+use crate::service::notifier_service::{Notifier, NotifierError};
+use crate::player::player::Player;
+use crate::player::track::PlaybackError;
 use crate::sources::spotify::spotify_client::{SpotifyClient, SpotifyError};
 use crate::sources::youtube::youtube_client::{SearchError, YoutubeClient};
 use dotenv::var;
