+++ src/commands/music/cmd_local.rs
@@ -3,7 +3,8 @@ use crate::checks::channel_checks::check_author_in_same_voice_channel;
 use crate::commands::music::cmd_download::{build_local_track, save_to_library, DownloadSource};
 use crate::embeds::music::player_embed::PlayerEmbed;
 use crate::embeds::music::queue_embed::QueueEmbed;
-use crate::player::player::{Player, Track};
+use crate::player::player::Player;
+use crate::player::track::Track;
 use crate::service::channel_service;
 use crate::service::embed_service::SendEmbed;
 use crate::service::picker_service::{self, PickerOutcome};
