+++ src/player/mod.rs
@@ -1,3 +1,3 @@
-pub mod notifier;
 #[allow(clippy::module_inception)]
 pub mod player;
+pub mod track;
