+++ src/service/mod.rs
@@ -4,4 +4,5 @@ pub mod channel_service;
 pub mod embed_service;
 pub mod gather_service;
 pub mod normalize_service;
+pub mod notifier_service;
 pub mod picker_service;
