+++ src/service/notifier_service.rs
